Refactor taxi type retrieval and improve nullable support
