+++ CabTap/CabTap.Contracts/Services/Identity/IUserService.cs
@@ -10,5 +10,5 @@ public interface IUserService
     Task<ClientDetailsViewModel> GetClientDetailsByIdAsync(string id);
     Task<ClientDeleteViewModel> GetClientToDeleteByIdAsync(string id);
     Task DeleteClientAsync(string id);
-    Task<string> GetUserId(string searchInput);
+    Task<string?> GetUserId(string searchInput);
 }
