+++ CabTap/CabTap.Contracts/Services/Utilities/ITaxiManagerService.cs
@@ -1,4 +1,5 @@
 using CabTap.Core.Entities.Enums;
+using CabTap.Shared.Category;
 using CabTap.Shared.Taxi;
 
 namespace CabTap.Contracts.Services.Utilities;
@@ -6,5 +7,6 @@ namespace CabTap.Contracts.Services.Utilities;
 public interface ITaxiManagerService
 {
     Task<TaxiAllViewModel> FindAvailableTaxiAsync(int categoryId);
+    Task<IEnumerable<CategoryPairViewModel>> GetAvailableTaxiTypesAsync();
     Task UpdateTaxiStatusAsync(int taxiId, TaxiStatus newStatus);
 }
