+++ CabTap/CabTap.Services/Services/Utilities/TaxiManagerService.cs
@@ -1,7 +1,9 @@
 using AutoMapper;
 using CabTap.Contracts.Repositories.Taxi;
+using CabTap.Contracts.Services.Taxi;
 using CabTap.Contracts.Services.Utilities;
 using CabTap.Core.Entities.Enums;
+using CabTap.Shared.Category;
 using CabTap.Shared.Taxi;
 using Microsoft.EntityFrameworkCore;
 
@@ -10,12 +12,14 @@ namespace CabTap.Services.Services.Utilities;
 public class TaxiManagerService : ITaxiManagerService
 {
     private readonly ITaxiRepository _taxiRepository;
+    private readonly ICategoryService _categoryService;
     private readonly IMapper _mapper;
 
-    public TaxiManagerService(ITaxiRepository taxiRepository, IMapper mapper)
+    public TaxiManagerService(ITaxiRepository taxiRepository, IMapper mapper, ICategoryService categoryService)
     {
         _taxiRepository = taxiRepository;
         _mapper = mapper;
+        _categoryService = categoryService;
     }
 
     public async Task<TaxiAllViewModel> FindAvailableTaxiAsync(int categoryId)
@@ -36,5 +40,34 @@ public class TaxiManagerService : ITaxiManagerService
         taxi.TaxiStatus = newStatus;
 
         await _taxiRepository.UpdateTaxiAsync(taxi);
+    }
+    
+    private static IEnumerable<CategoryPairViewModel> MapToCategoryPairViewModels(IEnumerable<CategoryPairViewModel> categories)
+    {
+        return categories.Select(x => new CategoryPairViewModel
+        {
+            Id = x.Id,
+            Name = x.Name,
+            Rate = x.Rate
+        });
+    }
+
+    public async Task<IEnumerable<CategoryPairViewModel>> GetAvailableTaxiTypesAsync()
+    {
+        var availableTaxis = _taxiRepository.GetTaxisQuery()
+            .Where(taxi => taxi.TaxiStatus == TaxiStatus.Available);
+
+        var allCategories = (await _categoryService.GetAllCategoriesAsync())
+            .ToList();
+        
+        var availableCategoryIds = availableTaxis
+            .Select(taxi => taxi.CategoryId)
+            .Distinct()
+            .ToHashSet();
+
+        var availableCategories = allCategories
+            .Where(category => availableCategoryIds.Contains(category.Id));
+
+        return MapToCategoryPairViewModels(availableCategories);
     }
 }
